LayoutTests/platform/ios-sim-deprecated/css3/device-adapt/opera/constrain-010-expected.txt


FAIL CSS Test: @viewport constrained - width 20rem assert_equals: expected 320 but got 800(stack: assert@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-010.xhtml:69:22
step@file:///Volumes/Data/WebKit.git/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/WebKit.git/LayoutTests/css3/device-adapt/opera/constrain-010.xhtml:68:14)
